file_storage/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
dj-database-url==2.0.0
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
django-storages==1.13.0
boto3==1.28.0
Pillow==10.0.1
psycopg2-binary==2.9.7